+++ pcs/cli/routing/cluster.py
@@ -50,6 +50,12 @@ cluster_cmd = create_router(
             {
                 "show": cluster.config_show,
                 "update": cluster.config_update,
+                "uuid": create_router(
+                    {
+                        "generate": cluster.generate_uuid,
+                    },
+                    ["cluster", "config", "uuid"],
+                ),
             },
             ["cluster", "config"],
             default_cmd="show",